// frontend/lib/api/generated/person/person.ts
    export type getApiPersonIdResponse200TextPlain = {
  data: Person
  status: 200
}

export type getApiPersonIdResponse200ApplicationJson = {
  data: Person
  status: 200
}

export type getApiPersonIdResponse200TextJson = {
  data: Person
  status: 200
}
// ...
export type getApiPersonIdResponseSuccess = (getApiPersonIdResponse200TextPlain | getApiPersonIdResponse200ApplicationJson | getApiPersonIdResponse200TextJson) & {
  headers: Headers;
};
// ...
export type putApiPersonIdResponse200TextPlain = {
  data: Person
  status: 200
}

export type putApiPersonIdResponse200ApplicationJson = {
  data: Person
  status: 200
}

export type putApiPersonIdResponse200TextJson = {
  data: Person
  status: 200
}
// ...
export type putApiPersonIdResponseSuccess = (putApiPersonIdResponse200TextPlain | putApiPersonIdResponse200ApplicationJson | putApiPersonIdResponse200TextJson) & {
  headers: Headers;
};